postgres fixed id generation

+++ pip-services4-postgres-node/src/persistence/IdentifiablePostgresPersistence.ts
@@ -153,6 +153,9 @@ export class IdentifiablePostgresPersistence<T extends IIdentifiable<K>, K> exte
      * @returns                 a found data item or <code>null</code>.
      */
     public async getOneById(context: IContext, id: K): Promise<T> {
+        if (this.isEmpty(id))
+            return null;
+        
         const query = "SELECT * FROM " + this.quotedTableName() + " WHERE \"id\"=$1";
         const params = [ id ];
 
@@ -192,7 +195,7 @@ export class IdentifiablePostgresPersistence<T extends IIdentifiable<K>, K> exte
 
         // Assign unique id
         let newItem: any = item;
-        if (newItem.id == null && this._autoGenerateId) {
+        if (this.isEmpty(newItem.id) && this._autoGenerateId) {
             newItem = Object.assign({}, newItem);
             newItem.id = item.id || IdGenerator.nextLong();
         }
@@ -214,7 +217,7 @@ export class IdentifiablePostgresPersistence<T extends IIdentifiable<K>, K> exte
         }
 
         // Assign unique id
-        if (item.id == null && this._autoGenerateId) {
+        if (this.isEmpty(item.id) && this._autoGenerateId) {
             item = Object.assign({}, item);
             item.id = <any>IdGenerator.nextLong();
         }
@@ -256,7 +259,7 @@ export class IdentifiablePostgresPersistence<T extends IIdentifiable<K>, K> exte
      * @returns                 the updated item.
      */
     public async update(context: IContext, item: T): Promise<T> {
-        if (item == null || item.id == null) {
+        if (item == null || this.isEmpty(item.id)) {
             return null;
         }
 
@@ -296,7 +299,7 @@ export class IdentifiablePostgresPersistence<T extends IIdentifiable<K>, K> exte
      * @returns                 the updated item.
      */
     public async updatePartially(context: IContext, id: K, data: AnyValueMap): Promise<T> {
-        if (data == null || id == null) {
+        if (data == null || this.isEmpty(id)) {
             return null;
         }
 
@@ -335,6 +338,9 @@ export class IdentifiablePostgresPersistence<T extends IIdentifiable<K>, K> exte
      * @returns                 the deleted item.
      */
     public async deleteById(context: IContext, id: K): Promise<T> {
+        if (this.isEmpty(id))
+            return null;
+
         const values = [ id ];
 
         const query = "DELETE FROM " + this.quotedTableName()
@@ -382,5 +388,21 @@ export class IdentifiablePostgresPersistence<T extends IIdentifiable<K>, K> exte
         });
 
         this._logger.trace(context, "Deleted %d items from %s", count, this._tableName);
+    }
+
+    /**
+     * Checks if value is empty
+     * @param value any value
+     * @returns true if value empty, other false
+     */
+    protected isEmpty(value: any) {
+        const type = typeof value;
+        if (value !== null && type === 'object' || type === 'function') {
+            const props = Object.keys(value);
+                if (props.length === 0) { 
+                    return true;
+                } 
+            } 
+        return !value;
     }
 }
